requirements.txt
# API y Autenticación
google-api-python-client==2.108.0
oauth2client==3.0.0  # Añadir esta versión específica
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# RSS Feed
feedparser==6.0.10

# Utilidades
requests==2.31.0
colorama==0.4.6
python-dateutil==2.8.2
pytz==2023.3.post1
python-telegram-bot==13.7

# Nota: sqlite3 ya está incluido en Python estándar, no necesita instalación
